requirements.txt
streamlit==1.37.1
chromadb==0.5.5
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.0
google-generativeai==0.7.2
openai==1.44.0
markdownify==0.12.1
urllib3==2.2.2
psutil==5.9.8
memory-profiler==0.61.0
langchain-text-splitters==0.2.2
